small fixes

+++ src/nevu_ui/fast/nvvector2/nvvector2.pxd
@@ -5,11 +5,11 @@ cdef class NvVector2:
     cdef public double y
     
     @staticmethod
-    cdef NvVector2 new(float x, float y)
+    cdef NvVector2 new(double x, double y)
 
     cdef NvVector2 _add(self, NvVector2 other)
     cdef NvVector2 _sub(self, NvVector2 other)
-    cdef NvVector2 _mul_scalar(self, float val)
+    cdef NvVector2 _mul_scalar(self, double val)
     cdef NvVector2 _mul_vector(self, NvVector2 other)
     cdef NvVector2 _iadd(self, NvVector2 other)
     cdef NvVector2 _isub(self, NvVector2 other)
